% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{basis_matrix}
\alias{basis_matrix}
\title{Evaluate a spline basis matrix at the specified times}
\usage{
basis_matrix(times, basis, integrate = FALSE)
}
\arguments{
\item{basis}{Info on the spline basis.}

\item{integrate}{A logical, should the integral of the basis be returned?}

\item{time}{A numeric vector.}
}
\value{
A two-dimensional array.
}
\description{
Evaluate a spline basis matrix at the specified times
}
