% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{calculate_log_crude_event_rate}
\alias{calculate_log_crude_event_rate}
\title{Calculate log_crude_event_rate
Helps to centre the intercept and the linear predictor}
\usage{
calculate_log_crude_event_rate(xdata, time_var = "time", status_var = "status")
}
\arguments{
\item{xdata}{dataset with time and status variables.}
}
\description{
Calculate log_crude_event_rate
Helps to centre the intercept and the linear predictor
}
