% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_cmdstan.R
\name{check_complete_post}
\alias{check_complete_post}
\title{Check posterior completeness}
\usage{
check_complete_post(post, ndraws)
}
\arguments{
\item{post}{A data.frame of posterior draws.}

\item{ndraws}{Expected total number of draws.}
}
\value{
Logical \code{TRUE} if counts match; otherwise an error.
}
\description{
Asserts that a posterior data.frame has the expected number of rows.
}
