% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGridOptimizeR.R
\name{cleanup_model}
\alias{cleanup_model}
\title{Clean up compiled model files}
\usage{
cleanup_model(modelName, modelDir)
}
\arguments{
\item{modelName}{Name of the model.}

\item{modelDir}{Path to directory containing the model binary.}
}
\description{
Removes the compiled model binary and .hpp file.
}
