% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGridOptimizeR.R
\name{clear_csv_model}
\alias{clear_csv_model}
\title{Remove CSV output files}
\usage{
clear_csv_model(modelName, modelDir)
}
\arguments{
\item{modelName}{Name of the model.}

\item{modelDir}{Path to directory containing the CSV files.}
}
\description{
Removes all CSV output files from a model directory.
}
