% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanTools.R
\name{colVars}
\alias{colVars}
\title{Column-wise variance}
\usage{
colVars(a)
}
\arguments{
\item{a}{A numeric matrix.}
}
\value{
A numeric vector of column variances.
}
\description{
Computes variance for each column of a matrix.
}
