% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_skeleton.R
\name{create_project_skeleton}
\alias{create_project_skeleton}
\title{Create project directory skeleton}
\usage{
create_project_skeleton(
  path = ".",
  dirs = c("Models", "Scripts", "DerivedData", "SourceData", "Simulations", "Img",
    "Results")
)
}
\arguments{
\item{path}{Path to project directory (default current directory).}

\item{dirs}{Character vector of directory names to create.}
}
\description{
Creates a standard project directory structure for Stan modelling workflows.
}
