% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{edit_model}
\alias{edit_model}
\title{Open a Stan model file in the editor}
\usage{
edit_model(mod_name)
}
\arguments{
\item{mod_name}{Name of the model (corresponds to directory name under Models/).}
}
\description{
Open a Stan model file in the editor
}
