% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{get_basis}
\alias{get_basis}
\title{Return the desired spline basis for the given knot locations}
\usage{
get_basis(
  x,
  iknots,
  bknots = range(x),
  degree = 3,
  intercept = TRUE,
  type = c("bs", "is", "ms")
)
}
\description{
Return the desired spline basis for the given knot locations
}
