% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{get_iknots}
\alias{get_iknots}
\title{Return a vector with internal knots for 'x', based on evenly spaced quantiles}
\usage{
get_iknots(x, df = 6L, degree = 3L, iknots = NULL, intercept = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{df}{The degrees of freedom. If specified, then 'df - degree - intercept'.
knots are placed at evenly spaced percentiles of 'x'. If 'iknots' is
specified then 'df' is ignored.}
}
\value{
A numeric vector of internal knot locations, or NULL if there are
no internal knots.
}
\description{
Return a vector with internal knots for 'x', based on evenly spaced quantiles
}
