% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_to_array}
\alias{list_to_array}
\title{Convert a list of same-dimension arrays to a higher-dimensional array}
\usage{
list_to_array(x, name = NULL)
}
\arguments{
\item{x}{A list of numeric vectors, matrices, or arrays of equal dimensions.}

\item{name}{Optional name for error messages.}
}
\value{
An array with one additional dimension prepended.
}
\description{
Convert a list of same-dimension arrays to a higher-dimensional array
}
