% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nlist}
\alias{nlist}
\title{Create a named list}
\usage{
nlist(...)
}
\arguments{
\item{...}{Objects to include in the list.}
}
\value{
A named list.
}
\description{
Creates a named list where names are automatically inferred from
the variable names of the arguments.
}
