% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padding.R
\name{pad}
\alias{pad}
\title{Pad an Object with NAs}
\usage{
pad(x, n, pad_numeric = 0)
}
\arguments{
\item{x}{An \R object (list, data.frame, matrix, atomic vector).}

\item{n}{The final length of each object.}

\item{pad_numeric}{handy if want to pad with something different than 0.}
}
\description{
This function pads an \R object (list, data.frame, matrix, atomic vector)
with \code{0}s. For matrices, lists and data.frames, this occurs by extending
each (column) vector in the object.
}
