% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parameter_table}
\alias{parameter_table}
\title{Parameter summary table}
\usage{
parameter_table(THETAhat, THETAomega)
}
\arguments{
\item{THETAhat}{Draws for fixed-effect parameters.}

\item{THETAomega}{Draws for random-effect parameters.}
}
\value{
A data.frame with Parameter, Estimate, RSE, and IIV columns.
}
\description{
Summarises posterior estimates with RSE and IIV.
}
