% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{preprocess_basis}
\alias{preprocess_basis}
\title{High-level function to preprocess basis
Assumes right censoring setting}
\usage{
preprocess_basis(time, status, min_time = 0, degree = 3, df = 6)
}
\arguments{
\item{time}{observed times.}

\item{status}{status}

\item{min_time}{lower bound time defult to 0.}

\item{degree}{splines degree default to 3.}

\item{df}{degrees of freedom default to 6.}
}
\description{
High-level function to preprocess basis
Assumes right censoring setting
}
\examples{
# not run
# rstanarm:::make_basis
}
