% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesGridOptimizeR.R
\name{runDiagnose}
\alias{runDiagnose}
\title{Run CmdStan diagnose}
\usage{
runDiagnose(model, data, init, seed, chain = 1, refresh = 100)
}
\arguments{
\item{model}{Path to model directory.}

\item{data}{Path to data file.}

\item{init}{Path to init file.}

\item{seed}{Random seed.}

\item{chain}{Chain ID.}

\item{refresh}{Refresh rate for output.}
}
\description{
Runs the CmdStan \code{diagnose} method on a compiled model.
}
