% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanTools.R
\name{summary.mcmc.list}
\alias{summary.mcmc.list}
\title{Summary method for mcmc.list}
\usage{
\method{summary}{mcmc.list}(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{An mcmc.list object.}

\item{quantiles}{Numeric vector of quantiles to compute.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An object of class \code{summary.mcmc}.
}
\description{
Custom summary for mcmc.list objects producing mean, SD, quantiles,
and time-series SE.
}
