% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{theoph}
\alias{theoph}
\title{PK of 12 subjects theophiline}
\format{
A data frame with 132 rows and 5 variables:
\describe{
\item{Subjec}{Subject id}
\item{conc}{concentration}
...
}
}
\source{
\url{https://raw.githubusercontent.com/tsahota/NMproject/master/inst/extdata/examples/theopp/SourceData/THEOPP.csv?_sm_au_=iVVNZMttDn4v60R2pGsWvKttvN1NG}
}
\usage{
theoph
}
\description{
A dataset containing the theophiline PK and other attributes of 12 subjects
}
\keyword{datasets}
