% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_stan_json}
\alias{write_stan_json}
\title{Write data in Stan JSON format}
\usage{
write_stan_json(data, file)
}
\arguments{
\item{data}{A named list of R objects (numeric vectors, matrices, data.frames, logicals, or lists).}

\item{file}{Path to the output JSON file.}
}
\description{
Writes a named list of data to a JSON file suitable for CmdStan input.
}
